refactor: CurrencyApi endpoints in currency.py
