+++ data_bridges_knots/client.py
@@ -20,6 +20,8 @@ from data_bridges_knots.endpoints.household import (
     get_household_surveys_list,
     get_household_xlsform_definition,
 )
+from data_bridges_knots.endpoints.currency import get_exchange_rates
+
 from data_bridges_knots.helpers import get_adm0_code
 
 logname = "data_bridges_api_calls.log"
@@ -209,68 +211,7 @@ class DataBridgesKnots:
         logger.debug("Token used: %s", token.__repr__())
         return configuration
 
-    def get_exchange_rates(
-        self, country_iso3: str, page_size: int = 1000
-    ) -> pd.DataFrame:
-        """Retrieves exchange rates for a given country from the Data Bridges API.
-
-        Args:
-            country_iso3 (str): The ISO3 country code
-            page_size (int, optional): Number of items per page. Defaults to 1000
-
-        Returns:
-            pd.DataFrame: DataFrame containing exchange rate data with columns:
-                - date: Date of exchange rate
-                - rate: Exchange rate value
-                - currency: Currency code
-                And other relevant exchange rate information
-
-        Examples:
-            >>> client = DataBridgesShapes("data_bridges_api_config.yaml")
-            >>> # Get exchange rates for Ethiopia
-            >>> rates_df = client.get_exchange_rates("ETH")
-            >>> # Check latest exchange rate
-            >>> latest_rate = rates_df.sort_values('date').iloc[-1]
-
-        Raises:
-            ApiException: If there's an error calling the Exchange rates API
-        """
-
-        responses = []
-        total_items = 20
-        max_item = 0
-        page = 0
-        while total_items > max_item:
-            page += 1
-            with data_bridges_client.ApiClient(
-                self._setup_configuration_and_authentication(self.config)
-            ) as api_client:
-                api_instance = data_bridges_client.CurrencyApi(api_client)
-                env = self.env
-
-                try:
-                    api_exchange_rates = (
-                        api_instance.currency_usd_indirect_quotation_get(
-                            country_iso3=country_iso3, format="json", page=page, env=env
-                        )
-                    )
-                    responses.extend(
-                        item.to_dict() for item in api_exchange_rates.items
-                    )
-                    total_items = api_exchange_rates.total_items
-                    logger.info("Fetching page %s", page)
-                    max_item = page * page_size
-                    time.sleep(1)
-                except ApiException as e:
-                    logger.error(
-                        "Exception when calling Exchange rates data-> : %s\n",
-                        e,
-                    )
-                    raise
-        df = pd.DataFrame(responses)
-        df = df.replace({np.nan: None})
-        return df
-
+ 
     def get_commodities_list(
         self,
         country_iso3: Optional[str] = None,
@@ -1418,11 +1359,12 @@ class DataBridgesShapes(DataBridgesKnots):
             f"Brought to you with <3 by WFP VAM"
         )
 
-
 DataBridgesKnots.get_household_survey = get_household_survey
 DataBridgesKnots.get_household_surveys_list = get_household_surveys_list
 DataBridgesKnots.get_household_xlsform_definition = get_household_xlsform_definition
 DataBridgesKnots.get_household_questionnaire = get_household_questionnaire
+DataBridgesKnots.get_exchange_rate = get_exchange_rates
+
 
 if __name__ == "__main__":
     pass
